+++ RELEASE_NOTES.md
@@ -36,6 +36,9 @@ no longer check that the list arguments are lists or partial lists. This is
 in line with current practice in Prolog compilers and draft standardization
 proposals where such checks are specified as optional.
 
+* CHANGED: Turning on the `debug` compiler flag no longer turns on the `clean`
+compiler flag.
+
 * IMPROVED: Performance of local meta-calls when using the `call/2-N` built-in
 predicates and the `call//1-N` built-in non-terminals with a callable closure
 known at compile time.